Add `get_serializer_model_instance` method
The `NestedModelViewSetMixin.get_serializer_model_instance` method
returns the unsaved instance of this viewsets model to use on
`create`.  The nesting model instance is set in this unsaved instance
and will be merged with the serializer's data on save.  The nesting
model instance is passed to the serializer via this unsaved instance
rather than via the `serializer.save()` call (e.g.
`serializer.save(nesting_model=model_inst)`) to make the nesting model
available in the serializer during validation, which is not the case
with the `serializer.save()` method.